+++ lib/read/cardano-wallet-read.cabal
@@ -61,6 +61,7 @@ library
     Cardano.Read.Ledger.Tx.Mint
     Cardano.Read.Ledger.Tx.ReferenceInputs
     Cardano.Read.Ledger.Tx.ScriptValidity
+    Cardano.Read.Ledger.Tx.Validity
     Cardano.Wallet.Read
     Cardano.Wallet.Read.Block
     Cardano.Wallet.Read.Block.BHeader
@@ -95,7 +96,6 @@ library
     Cardano.Wallet.Read.Tx.Inputs
     Cardano.Wallet.Read.Tx.TxId
     Cardano.Wallet.Read.Tx.TxIn
-    Cardano.Wallet.Read.Tx.Validity
     Cardano.Wallet.Read.Tx.Withdrawals
     Cardano.Wallet.Read.Tx.Witnesses
